backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6
django-cors-headers==4.9.0
